# Avoid copying buffer from ReadOnlySequence as  ReceivedMqttPacket body