--- a/Source/MQTTnet.Benchmarks/Program.cs
+++ b/Source/MQTTnet.Benchmarks/Program.cs
@@ -31,7 +31,7 @@ namespace MQTTnet.Benchmarks
             Console.WriteLine("e = MessageDeliveryBenchmark");
             Console.WriteLine("f = AsyncLockBenchmark");
             Console.WriteLine("g = MqttBufferReaderBenchmark");
-            Console.WriteLine("g = ReaderExtensionsBenchmark");
+            Console.WriteLine("h = ReaderExtensionsBenchmark");
 
             var pressedKey = Console.ReadKey(true);
             switch (pressedKey.KeyChar)
@@ -80,6 +80,12 @@ namespace MQTTnet.Benchmarks
                     break;
                 case 'f':
                     BenchmarkRunner.Run<AsyncLockBenchmark>();
+                    break;
+                case 'g':
+                    BenchmarkRunner.Run<MqttBufferReaderBenchmark>();
+                    break;
+                case 'h':
+                    BenchmarkRunner.Run<ReaderExtensionsBenchmark>();
                     break;
             }