### Related project
MQTTnet.AspNetCore

### How it affects
Read `ArraySegment<byte>` buffer from ReadOnlySequence as ReceivedMqttPacket body. 
It can greatly reduce unnecessary buffer-copying.

### Dependency
This request depend on #1592，after we make MqttBufferReader correctly support `ArraySegment<byte>`