@xljiulang 2022-11-29
The benchmark result shows a good reduction in memory allocation!
``` ini

BenchmarkDotNet=v0.13.2, OS=Windows 10 (10.0.19044.2251/21H2/November2021Update)
Intel Core i3-4150 CPU 3.50GHz (Haswell), 1 CPU, 4 logical and 2 physical cores
.NET SDK=7.0.100
  [Host]   : .NET 6.0.8 (6.0.822.36306), X64 RyuJIT AVX2
  .NET 6.0 : .NET 6.0.8 (6.0.822.36306), X64 RyuJIT AVX2

Job=.NET 6.0  Runtime=.NET 6.0  

```
| Method |     Mean |     Error |    StdDev | Ratio | Rank |    Gen0 | Allocated | Alloc Ratio |
|------- |---------:|----------:|----------:|------:|-----:|--------:|----------:|------------:|
| Before | 3.593 μs | 0.0124 μs | 0.0097 μs |  1.00 |    2 | 13.4201 |  20.57 KB |        1.00 |
|  After | 2.536 μs | 0.0151 μs | 0.0134 μs |  0.71 |    1 |  6.8932 |  10.54 KB |        0.51 |